+++ CHANGELOG.md
@@ -54,6 +54,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fundamentals/prototype for wiring system
 - Mountain peak and lake markers on the map
 - There's now a checkbox in the graphics tab to opt-in to receiving lossily-compressed terrain colors.
+- /buff command which allows you to cast a buff on player
 
 ### Changed